requirements.txt
flet==0.27.6
pyrebase4==4.7.1
requests==2.31.0